updated welcome modal with consent changes
