+++ src/screens/application/children/welcomeModal.tsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import { useIntl } from 'react-intl';
-import { Text, Image, View, SafeAreaView, TouchableOpacity } from 'react-native';
+import { Text, Image, View, SafeAreaView, TouchableOpacity, Linking } from 'react-native';
 
 import EStyleSheet from 'react-native-extended-stylesheet';
 import VersionNumber from 'react-native-version-number';
 
-import { Icon, MainButton, Modal } from '../../../components';
+import { CheckBox, Icon, MainButton, Modal } from '../../../components';
 import { useAppDispatch, useAppSelector } from '../../../hooks';
 import { setLastAppVersion } from '../../../redux/actions/applicationActions';
 import parseVersionNumber from '../../../utils/parseVersionNumber';
@@ -21,11 +21,15 @@ const WelcomeModal = ({ onModalVisibilityChange }) => {
 
   const [showAnimation, setShowAnimation] = useState(true);
   const [showWelcomeModal, setShowWelcomeModal] = useState(false);
+  const [isConsentChecked, setIsConsentChecked] = useState(false);
+
   const [appVersion] = useState(VersionNumber.appVersion);
 
 
   useEffect(() => {
     _compareAppVersion()
+    _showWelcomeModal();
+
   }, [])
 
 
@@ -51,6 +55,10 @@ const WelcomeModal = ({ onModalVisibilityChange }) => {
     onModalVisibilityChange(false);
   }
 
+  const _onCheckPress = (value, isCheck) => {
+    setIsConsentChecked(isCheck);
+  };
+
 
   const _renderInfo = (iconName, headingIntlId, bodyIntlId) => (
     <View style={styles.sectionRow}>
@@ -64,6 +72,26 @@ const WelcomeModal = ({ onModalVisibilityChange }) => {
         <Text style={styles.sectionTitle}>{intl.formatMessage({ id: headingIntlId })}</Text>
         <Text style={styles.sectionText}>{intl.formatMessage({ id: bodyIntlId })}</Text>
       </View>
+    </View>
+  );
+
+  const _renderConsent = () => (
+    <View style={styles.consentContainer}>
+      <CheckBox value={isConsentChecked} clicked={_onCheckPress} style={styles.checkStyle} />
+      <TouchableOpacity disabled={!isConsentChecked} onPress={() => Linking.openURL('https://ecency.com/terms-of-service')}>
+        <View style={styles.consentTextContainer}>
+          <Text style={styles.termsDescText}>
+            {intl.formatMessage({
+              id: 'welcome.terms_description',
+            })}
+            <Text style={styles.termsLinkText}>
+              {intl.formatMessage({
+                id: 'welcome.terms_text',
+              })}
+            </Text>
+          </Text>
+        </View>
+      </TouchableOpacity>
     </View>
   );
 
@@ -83,11 +111,12 @@ const WelcomeModal = ({ onModalVisibilityChange }) => {
           {_renderInfo('question', 'welcome.line1_heading', 'welcome.line1_body')}
           {_renderInfo('emotsmile', 'welcome.line2_heading', 'welcome.line2_body')}
           {_renderInfo('people', 'welcome.line3_heading', 'welcome.line3_body')}
+          {_renderConsent()}
         </View>
         <View style={styles.bottomButton}>
           <MainButton
             onPress={_handleButtonPress}
-            isDisable={false}
+            isDisable={!isConsentChecked}
             isLoading={false}
             style={{ alignSelf: 'center', paddingHorizontal: 30 }}
             text={intl.formatMessage({ id: 'welcome.get_started' })}
